Add String.hpp header and unit tests
This change adds a new file to the library: String.hpp provides routines
for simple interaction with C-style character arrays. The two currently
supported functions are: length() for determining the length of a
string and compare() for lexical comparison of two strings. Unit tests
are added along with the actual functionality.