Add copy() function for copying C-style strings
This change adds a copy() function for creating a copy of a zero
terminated C-style string. Like strncpy, the function requires users to
pass in the capacity of the destination buffer. Unlike strncpy, however,
the resulting string is always zero terminated. At the same time, the
remaining elements of the destination buffer are never cleared if the
source string is shorter than the destination buffer.