Fix bug in includes() function
The includes() function used for checking whether an iterator is
included in a range specified by two iterators does not work for
wrapping iterators such as libcontainer's MoveIterator. The reason is
that we used the iterators' types to aid that decision at compile time:
if the types of the iterator to check ('it') and the iterators spanning
the range of interest ('range') point to different types the assumption
is that 'it' can never point to an element in 'range'.
This is generally true, however, the implementation did not take care
of the case that a wrapping iterator is used that just wraps another
iterator but effectively represents it. In this case, includes() reports
wrong results (returns false).
This change fixes the problem by actually evaluating the type that
operator *() of the iterators of interest yield. The basic
implementation and principles stay the same.
In addition, a unit test is added that tests 'includes' for "normal"
usage and in conjunction with wrapping iterators.