Declare rotateLeft() and rotateRight() functions as constexpr
The two functions rotateLeft() and rotateRight() can also be declared a
constexpr since the functionality can be expressed within a single
return statement.
This change makes declares both functions as constexpr. Note that using
constexpr we can no longer assert for preconditions in the function
body. The assert statements are commented out but left in for
documentation purposes.