Import Assert.hpp header from libelfload
For the development of the ELF loader a custom implementation of
assertions was implemented in the form of the ASSERT macro. This change
imports the Assert.hpp header containing the definition to make it more
widely usable.
The assertion implementation is more sophisticated than the assert()
implementation provided by standard system headers in that it provides a
version that accepts a operation to perform on two arguments and, if the
result fails to be true, prints the actual values of the arguments
passed in. Such behavior makes it more easy to investigate problems
without actually attaching a debugger or performing similar means.