Rename ASSERT macro to ASSERT__
Currently, the ASSERT macro as provided by Assert.hpp clashes with ASSERT from
libtest which is only used for unit testing. The plan is to rename libtest's
ASSERT to something different while keeping libutil's version of ASSERT.
To ease the transition this change temporarily renames ASSERT to ASSERT__. This
way it is easy to identify all clients that use the actual ASSERT of libtest.