Use RemoveReference for type deduction in ASSERTOP_IMPL
ASSERTOP_IMPL uses decltype to infer the types of their arguments such
that an operation object can be instantiated.
However, if the passed-in types are references, instantiation will fail
because pointers (which are created based on these inferred types)
cannot point to references. This change uses typ::RemoveReference to
avoid this problem.