Do not require pointer to function in assertOp
Until now 'assertOp' accepted a pointer to a function matching the
signature of 'AssertFailFunction'. This approach does not work in
conjunction with lambda expressions and, thus, is of limited
flexibility.
This change converts the last parameter of 'assertOp' to be deducible by
template instantiation. This means that from now on lambda expressions
and functors can be accepted by this function as well.