Adhere to ASSERT related renaming in libtest
The ASSERT macro used by libtest clashes with ASSERT provided by
libutil. The latest change to libtest renamed its ASSERT macro to
TESTASSERT. Along with this change it also based its implementation on
libutil's assertion infrastructure. Because of that, a more
sophisticated TESTASSERTOP macro was introduced as well that allows for
easier debugging in the case of assertion failures since the values of
the parameters of the failed assertion are printed.
This change adjusts all unit tests to adhere to the renaming and makes
use of the new TESTASSERTOP macro where possible.