Fix compile error when deduced type is const
With a new version of gcc, when OutStream::printUnsignedValue might be invoked
with a type T that is const. In that case, OutStream::printUnsignedValueImpl
deduce the type T as const as well which leads to compile errors due to
assignment to constant variables.
This change fixes this issue by explictly removing the constness from the type.