Add exception specifier to __assert_fail
When directly or indirectly including cassert, the C++ standard header
for assertions, one implicitly pulls in a declaration of __assert_fail
that contains and empty exception specifier "throw()". Unfortunately,
this declaration conflicts with the declaration provided by this
library.
This patch adjusts this library's declaration to contain the empty
exception specifier as well. Note that long term the usage of cassert is
discouraged but in the interim (and even once nobody uses the standard
header anymore), having this exception specifier has no foreseeable
negative impact.